remove lighttypes
