--- examples/lighting-app/silabs/include/LightTypes.h
@@ -1,46 +1,0 @@
-/*
- *
- *    Copyright (c) 2020 Project CHIP Authors
- *    Copyright (c) 2019 Google LLC.
- *    All rights reserved.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-#pragma once
-
-#include <stdint.h>
-
-namespace LightingManager {
-
-enum Action_t
-{
-    ON_ACTION = 0,
-    OFF_ACTION,
-    LEVEL_ACTION,
-    COLOR_ACTION_HSV,
-    COLOR_ACTION_CT,
-    COLOR_ACTION_XY,
-
-    INVALID_ACTION
-};
-
-enum State_t
-{
-    kState_OffInitiated = 0,
-    kState_OffCompleted,
-    kState_OnInitiated,
-    kState_OnCompleted,
-};
-
-} // namespace LightingManager
