+++ examples/lighting-app/silabs/src/DataModelCallbacks.cpp
@@ -25,7 +25,6 @@
 #include "RGBLEDWidget.h"
 #endif //(defined(SL_MATTER_RGB_LED_ENABLED) && SL_MATTER_RGB_LED_ENABLED == 1)
 #include "AppTask.h"
-#include "LightTypes.h"
 
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app-common/zap-generated/ids/Attributes.h>
